6.Project/requirements.txt
falcon
markdown
PyJWT
waitress